frontend: reactive UI and auto-reconnect
So I wanted to achieve automatic discovery of hierarchically dependent values that are transformed (rendered) to Virtual DOM objects. So, algorithm turned out to be simple. If some value is updated, it's marked as invalidated, no matter who updates it. When it's transformed (mapped/rendered) to then it gets the render session which is just a stack of current render tree. If it wasn't previously invalidated then it just returns the cache and the stack is not altered. Otherwise, if the value's mapping (the transform) needs to be recalculated because value changed then it means that all its ascendants (but not neighbours of ascendants) from the stack need to be invalidated. All invalidations cause ditching cached value transforms where the invalidated values are used. The last bit is to call root re-render function that just uses the transform mechanism and all it's caching goodness.

Because this is not a compiled DSL, I have to use some utils (classes, functions) to achieve all that. For values that are updated and are transformed to render results, I have a ValueContainer<T> which has got a method `renderTo(r: RenderSession, fn: (TData) -> TCachedResult)` which creates an object caching last result and last value's version ID. When it's called then it compares the current version of a value and it's own cache ID, then decides whether transform should happen again or just read from the cache.