WIP watchable value container, transforming multiple with tree-dependent changes
This got actually too complicated. I tried to not entangle data for certain views
but it's pretty hard to detect changes and inform enough amount of visual tree to update.
So, I guess it will be easier to manage if data was bound to a view function.
Component architecture may come in.